Create String_add.cpp
A very simple implementation of representing positive integers as strings in C++, and performing multiplication on two such string-represented integers.  The output of the various functions has been described briefly in the comments.  The logic is pretty simple: any two integers, each less than 9999, can be multiplied in the trivial way as C++ int's.  Anything larger is multiplied using recursion by using the well-known formula: x*y=z_2*10^(2m)+z_1*10^m+z_0, where z_2, z_1, and z_0 are given by: x_1*y_1, x_1*y_0+x_0*y_1, and x_0*_y_0, for a representation x=x_1*10^m+x_0, y=y_1*10^m+y_0.  After writing a function to subtract such strings, one can easily implement Karatsuba's famous multiplication algorithm, but I have not done this yet.  

This code compiles with the single warning of the C++ compiler not being able to locate a "return" statement outside an "if" clause in the function Mult(a,b).  My version of g++ is 1300.0.29.30.  If this sort of thing annoys you, you can insert a return "hello" on line 176, which will never be reached in the execution of the code.